+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Removed
 
 - **Projectroles**
     - Support for deprecated search results as dict (#1400)
+    - Support for deprecated app setting ``local`` parameter (#1394)
 - **Timeline**
     - Support for deprecated ``get_object_link()`` return values as dict (#1398)